+++ src/pages/MyPage.jsx
@@ -1,19 +1,10 @@
 import styled from 'styled-components';
-import sampleImg from '../assets/다운로드.jpeg';
 import { useContext } from 'react';
 import { AuthContext } from '../contexts/AuthProvider';
 import { useState } from 'react';
 import { supabase } from '../supabase/client';
 import { useEffect } from 'react';
-
-const dummyArr = Array.from({ length: 4 }, (_, idx) => ({
-  post_num: idx + 1,
-  post_title: '춥고 배고프고 졸려',
-  post_date: '1 hours ago',
-  post_img: sampleImg,
-  post_like: 1,
-  user_id: '임재원',
-}));
+import MyPostList from '../components/MyPage/MyPostList';
 
 const MyPage = () => {
   const { user, setUser } = useContext(AuthContext);
@@ -104,49 +95,32 @@ const MyPage = () => {
           ) : (
             <RoundButton></RoundButton>
           )}
-          {isEditing ? <input type="file" onChange={handleFileChange} /> : ''}
+          {isEditing && <input type="file" onChange={handleFileChange} />}
         </StProfileWrapper>
         <StMyInfoWrapper>
           <StNickname>{user.nickname}</StNickname>
-          {isEditing ? (
+          {isEditing && (
             <input
               type="text"
               value={editedNickname}
               onChange={(e) => setEditedNickname(e.target.value)}
             />
-          ) : (
-            ''
           )}
           <StIntroduce>{user.intro}</StIntroduce>
-          {isEditing ? (
+          {isEditing && (
             <textarea
               name=""
               id=""
               value={editedIntro}
               onChange={(e) => setEditedIntro(e.target.value)}
             ></textarea>
-          ) : (
-            ''
           )}
         </StMyInfoWrapper>
         <StEditBtn value={isEditing} onClick={handleUserInfoChange}>
           내 정보 수정하기
         </StEditBtn>
       </StMyInfoChange>
-      <StMyPostList>
-        {dummyArr.map((post) => {
-          return (
-            <CardContainer key={post.post_num}>
-              <PostImage>이미지</PostImage>
-              <MyPostWrapper>
-                <MyPostTitle>제목</MyPostTitle>
-                <div>날짜</div>
-                <div>좋아요</div>
-              </MyPostWrapper>
-            </CardContainer>
-          );
-        })}
-      </StMyPostList>
+      <MyPostList></MyPostList>
     </div>
   );
 };
@@ -162,39 +136,6 @@ const RoundImage = styled.img`
 const StProfileWrapper = styled.div`
   display: flex;
   flex-direction: column;
-`;
-
-const MyPostWrapper = styled.div`
-  margin: 10px;
-`;
-
-const MyPostTitle = styled.div`
-  font-size: 30px;
-`;
-
-const StMyPostList = styled.div`
-  background-color: #fee3a2;
-  height: 500px;
-  border-radius: 30px;
-  margin: 40px auto;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  padding: 80px;
-`;
-
-const CardContainer = styled.div`
-  width: 450px;
-  background-color: #b4972b;
-  height: 150px;
-  border-radius: 5px;
-  display: flex;
-`;
-const PostImage = styled.div`
-  width: 150px;
-  height: 150px;
-  background-color: white;
-  border-radius: 5px;
 `;
 
 const StMyInfoChange = styled.div`
